services/worker/requirements.txt
redis==5.0.8
rq==1.16.0
boto3==1.35.18
requests==2.32.5
Pillow==10.4.0
imagehash==4.3.1
loguru==0.7.2
python-dotenv==1.0.1
pydantic==2.9.2